+++ packages/magic-scripts/config/webpack.config.prod.js
@@ -153,6 +153,19 @@ module.exports = {
         loader: ExtractTextPlugin.extract('style', 'css?importLoaders=1&-autoprefixer!postcss')
         // Note: this won't work without `new ExtractTextPlugin()` in `plugins`.
       },
+      {
+        test: /\.less$/,
+        // "?-autoprefixer" disables autoprefixer in css-loader itself:
+        // https://github.com/webpack/css-loader/issues/281
+        // We already have it thanks to postcss. We only pass this flag in
+        // production because "css" loader only enables autoprefixer-powered
+        // removal of unnecessary prefixes when Uglify plugin is enabled.
+        // Webpack 1.x uses Uglify plugin as a signal to minify *all* the assets
+        // including CSS. This is confusing and will be removed in Webpack 2:
+        // https://github.com/webpack/webpack/issues/283
+        loader: ExtractTextPlugin.extract('style', 'css?importLoaders=1&-autoprefixer!postcss!less')
+        // Note: this won't work without `new ExtractTextPlugin()` in `plugins`.
+      },
       // JSON is not enabled by default in Webpack but both Node and Browserify
       // allow it implicitly so we also enable it.
       {